兼容服务器

+++ MPI/FlatTransform.h
@@ -7,11 +7,11 @@ namespace FlatPhysics {
 	public:
 		FlatTransform() :position({}),sin(0),cos(0) {}
 		FlatTransform(const Vector2& position, float angle)
-			:position(position),sin(std::sinf(angle)),cos(std::cosf(angle))
+			:position(position),sin(std::sin(angle)),cos(std::cos(angle))
 		{
 		}
 		FlatTransform(float x, float y, float angle)
-			:position({x,y}), sin(std::sinf(angle)), cos(std::cosf(angle))
+			:position({x,y}), sin(std::sin(angle)), cos(std::cos(angle))
 		{
 		}
 		static FlatTransform Zero() { return FlatTransform(0.0f, 0.0f,0.0f); }
